+++ gui/Document.cc
@@ -14,10 +14,10 @@
 
 // gui namespace headers
 #include "common/RaiiCursor.hh"
-#include "class_diagram/ClassModel.hh"
-#include "logical_view/LogicalModel.hh"
-#include "logical_view/ProxyModel.hh"
-#include "source_view/SourceModel.hh"
+#include "classgf/ClassModel.hh"
+#include "logicalvw/LogicalModel.hh"
+#include "logicalvw/ProxyModel.hh"
+#include "sourcevw/SourceModel.hh"
 
 #include "project/Project.hh"
 
@@ -31,6 +31,10 @@
 #include <iostream>
 
 namespace gui {
+
+using namespace logicalvw;
+using namespace sourcevw;
+using namespace classgf;
 
 namespace fs = boost::filesystem;
 
@@ -111,14 +115,14 @@ void Document::AddSource(const QString& file)
 
 void Document::NewClassDiagram(const QString& name)
 {
-	auto m = std::make_unique<gui::ClassModel>(&m_project->CodeBase().Map(), name, this);
+	auto m = std::make_unique<classgf::ClassModel>(&m_project->CodeBase().Map(), name, this);
 	emit OnCreateClassDiagramView(m.get());
 	m_project->Add(std::move(m));
 }
 
 void Document::NewSourceView(const QString& name, unsigned line, unsigned column)
 {
-	auto m = std::make_unique<gui::SourceModel>(name, this);
+	auto m = std::make_unique<sourcevw::SourceModel>(name, this);
 	emit OnCreateSourceView(m.get());
 	
 	m->SetLocation(name, line, column);
@@ -143,7 +147,7 @@ void Document::Open(const QString& file)
 	RaiiCursor cursor(Qt::WaitCursor);
 	ModelFactory factory{this};
 	proj->Open(file.toStdString(), factory);
-	
+
 	for (auto&& tu : proj->CodeBase().TranslationUnits())
 		for (auto&& diag : tu.Diagnostics())
 			emit OnCompileDiagnotics(QString::fromStdString(diag.Str()));
@@ -171,7 +175,7 @@ project::Model Document::ModelFactory::Create(project::ModelType type, const std
 	{
 	case project::ModelType::class_diagram:
 	{
-		auto m = std::make_unique<gui::ClassModel>(&owner.CodeBase().Map(), QString::fromStdString(name), m_parent);
+		auto m = std::make_unique<classgf::ClassModel>(&owner.CodeBase().Map(), QString::fromStdString(name), m_parent);
 		emit m_parent->OnCreateClassDiagramView(m.get());
 		result = std::move(m);
 		break;