+++ gui/classgf/Edge.cc
@@ -16,9 +16,9 @@
 
 #include <cassert>
 #include <cmath>
-#include <iostream>
 
 namespace gui {
+namespace classgf {
 
 const auto arrow_width = 15.0;
 
@@ -31,36 +31,40 @@ Edge::Edge(const BaseItem *from, const BaseItem *to) :
 	UpdatePosition();
 }
 
-void Edge::paint(QPainter *painter, const QStyleOptionGraphicsItem*, QWidget*)
+void Edge::paint(QPainter *painter, const QStyleOptionGraphicsItem *, QWidget *)
 {
 	auto from = m_from->GraphicsItem();
-	auto to   = m_to->GraphicsItem();
+	auto to = m_to->GraphicsItem();
 	
 	if (from->collidesWithItem(to))
 		return;
 	
 	QLineF dia{
 		mapFromItem(from, QPointF{from->boundingRect().center()}),
-		mapFromItem(to,   QPointF{to->boundingRect().center()})
+		mapFromItem(to, QPointF{to->boundingRect().center()})
 	};
 	
 	QPointF from_pt, to_pt;
-		
-	auto from_p = mapFromItem(from, from->boundingRect());
-	auto to_p   = mapFromItem(to,   to->boundingRect());
 	
-	for (int i = 0 ; i < from_p.size() ; ++i)
+	auto from_p = mapFromItem(from, from->boundingRect());
+	auto to_p = mapFromItem(to, to->boundingRect());
+	
+	for (int i = 0; i < from_p.size(); ++i)
 	{
-		QLineF line{from_p.at(i), from_p.at((i+1) % from_p.size())};
+		QLineF line{from_p.at(i), from_p.at((i + 1) % from_p.size())};
 		if (dia.intersect(line, &from_pt) == QLineF::BoundedIntersection)
 			break;
 	}
-	for (int i = 0 ; i < to_p.size() ; ++i)
+	for (int i = 0; i < to_p.size(); ++i)
 	{
-		QLineF line{to_p.at(i), to_p.at((i+1) % to_p.size())};
+		QLineF line{to_p.at(i), to_p.at((i + 1) % to_p.size())};
 		if (dia.intersect(line, &to_pt) == QLineF::BoundedIntersection)
 			break;
 	}
+	
+	QPen line_pen{Qt::GlobalColor::black};
+	line_pen.setCosmetic(true);
+	painter->setPen(line_pen);
 	
 	QLineF to_draw{from_pt, to_pt};
 	painter->drawLine(to_draw);
@@ -75,17 +79,17 @@ QRectF Edge::boundingRect() const
 void Edge::UpdatePosition()
 {
 	auto from = m_from->GraphicsItem();
-	auto to   = m_to->GraphicsItem();
+	auto to = m_to->GraphicsItem();
 	
 	auto from_center = from->boundingRect().center();
-	auto to_center   = to->boundingRect().center();
+	auto to_center = to->boundingRect().center();
 	
 	setPos(QRectF{from_center, to_center}.normalized().center());
 	
 	prepareGeometryChange();
 	m_bounding = QRectF{
 		mapFromItem(from, from_center),
-		mapFromItem(to,   to_center),
+		mapFromItem(to, to_center),
 	}.normalized().adjusted(-arrow_width, -arrow_width, arrow_width, arrow_width);
 }
 
@@ -94,7 +98,7 @@ ItemRelation Edge::RelationOf(const BaseItem *) const
 	return ItemRelation::no_relation;
 }
 
-gui::ItemType Edge::ItemType() const
+classgf::ItemType Edge::ItemType() const
 {
 	return ItemType::edge;
 }
@@ -102,14 +106,14 @@ gui::ItemType Edge::ItemType() const
 void Edge::DrawEndings(QPainter *painter, const QLineF& dia) const
 {
 	auto from_pt = dia.p1();
-	auto to_pt   = dia.p2();
+	auto to_pt = dia.p2();
 	
 	// build transform matrix for drawing at the point
 	QTransform head{painter->transform()}, tail{painter->transform()};
 	head.translate(from_pt.x(), from_pt.y());
 	tail.translate(to_pt.x(), to_pt.y());
 	
-	auto angle = -std::atan(dia.dx()/dia.dy());
+	auto angle = -std::atan(dia.dx() / dia.dy());
 	if (dia.dy() < 0)
 		angle += M_PI;
 	head.rotateRadians(angle);
@@ -134,9 +138,12 @@ void Edge::DrawToEnding(QPainter *painter, ItemRelation relation) const
 	
 	switch (relation)
 	{
-	case ItemRelation::base_class_of:           DrawInheritanceTrigangle(painter);  break;
-	case ItemRelation::use_as_member:           DrawArrowHead(painter);             break;
-	case ItemRelation::used_by_as_member:       DrawAggregationDiamond(painter);    break;
+	case ItemRelation::base_class_of: DrawInheritanceTrigangle(painter);
+		break;
+	case ItemRelation::use_as_member: DrawArrowHead(painter);
+		break;
+	case ItemRelation::used_by_as_member: DrawAggregationDiamond(painter);
+		break;
 	default: break;
 	}
 }
@@ -147,9 +154,12 @@ void Edge::DrawFromEnding(QPainter *painter, ItemRelation relation) const
 	
 	switch (relation)
 	{
-	case ItemRelation::derived_class_of:        DrawInheritanceTrigangle(painter);  break;
-	case ItemRelation::used_by_as_member:       DrawArrowHead(painter);             break;
-	case ItemRelation::use_as_member:           DrawAggregationDiamond(painter);    break;
+	case ItemRelation::derived_class_of: DrawInheritanceTrigangle(painter);
+		break;
+	case ItemRelation::used_by_as_member: DrawArrowHead(painter);
+		break;
+	case ItemRelation::use_as_member: DrawAggregationDiamond(painter);
+		break;
 	default: break;
 	}
 }
@@ -159,12 +169,12 @@ void Edge::DrawInheritanceTrigangle(QPainter *painter) const
 	assert(painter);
 	
 	painter->drawPolygon(
-		QPolygonF{} << QPointF{} << QPointF{arrow_width,arrow_width} << QPointF{-arrow_width, arrow_width},
+		QPolygonF{} << QPointF{} << QPointF{arrow_width, arrow_width} << QPointF{-arrow_width, arrow_width},
 		Qt::FillRule::WindingFill
 	);
 }
 
-void Edge::Update(const codebase::EntityMap*)
+void Edge::Update(const codebase::EntityMap *)
 {
 }
 
@@ -179,7 +189,7 @@ void Edge::DrawArrowHead(QPainter *painter) const
 {
 	assert(painter);
 	painter->drawPolyline(
-		QPolygonF{} << QPointF{arrow_width/2,arrow_width} << QPointF{} << QPointF{-arrow_width/2, arrow_width}
+		QPolygonF{} << QPointF{arrow_width / 2, arrow_width} << QPointF{} << QPointF{-arrow_width / 2, arrow_width}
 	);
 }
 
@@ -189,9 +199,9 @@ void Edge::DrawAggregationDiamond(QPainter *painter) const
 	painter->drawPolygon(
 		QPolygonF{}
 			<< QPointF{}
-			<< QPointF{arrow_width/2,arrow_width}
-			<< QPointF{0, arrow_width*2}
-			<< QPointF{-arrow_width/2,arrow_width},
+			<< QPointF{arrow_width / 2, arrow_width}
+			<< QPointF{0, arrow_width * 2}
+			<< QPointF{-arrow_width / 2, arrow_width},
 		Qt::FillRule::WindingFill
 	);
 }
@@ -206,4 +216,4 @@ const QGraphicsItem *Edge::GraphicsItem() const
 	return this;
 }
 	
-} // end of namespace
+}} // end of namespace