+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change Log
 
-## [Version 0.1](https://github.com/nestal/spaghetti/releases/tag/v0.1)
+## [Version 0.2](https://gitlab.com/nestal/spaghetti/builds) (Upcoming)
+
+- Support resizing classes in class diagram
+- Support aggregation relationship in class diagram
+- Show only used classes in logical view
+
+## [Version 0.1](https://github.com/nestal/spaghetti/releases/tag/v0.1) (25 February 2017)
 
 - Initial version
 - Support parsing C++ source files
 - Creating class diagram by drag-and-drop from logical view
+- Display inheritance relationship in class diagram
 - View source code by double-clicking in logical view